+++ relay-server/src/endpoints/playstation.rs
@@ -5,6 +5,7 @@ use axum::routing::{MethodRouter, post};
 use relay_config::Config;
 use relay_dynamic_config::Feature;
 use relay_event_schema::protocol::EventId;
+use serde::Serialize;
 
 use crate::endpoints::common::{self, BadStoreRequest, TextResponse};
 use crate::envelope::ContentType::OctetStream;
@@ -16,11 +17,48 @@ use crate::utils::UnconstrainedMultipart;
 /// The extension of a prosperodump in the multipart form-data upload.
 const PROSPERODUMP_EXTENSION: &str = ".prosperodmp";
 
-/// Prosperodump attachments should have these magic bytes
+/// Prosperodump attachments should have these magic bytes.
 const PROSPERODUMP_MAGIC_HEADER: &[u8] = b"\x7FELF";
 
 /// Magic bytes for lz4 compressed prosperodump containers.
 const LZ4_MAGIC_HEADER: &[u8] = b"\x04\x22\x4d\x18";
+
+/// Content type of the data request payload.
+const DATA_REQUEST_CONTENT_TYPE: &str = "application/vnd.sce.crs.datareq-request+json";
+
+/// Files that are currently supported for uploading.
+const UPLOAD_PARTS: &[&str] = &["corefile", "screenshot"];
+
+/// Implementation of the `getRequestCoreDumpData` [response](https://game.develop.playstation.net/resources/documents/SDK/11.000/Core_Dump_Data_Request_API-Reference/0002.html).
+///
+/// The optional `partParameters` section is omitted as we make no use of it. This is because any
+/// parameter specified there only applies to on demand generated videos. So it can't be used to
+/// truncate videos that are already present on the device.
+#[derive(Serialize)]
+struct DataRequestResponse {
+    parts: Parts,
+}
+
+/// Partial implementation of the `parts` section in the `getRequestCoreDumpData` response.
+///
+/// The optional `additional` and `no_upload` sections are currently omitted as they are not used.
+#[derive(Serialize)]
+struct Parts {
+    /// List of files that can be uploaded.
+    ///
+    /// Note: It is not guaranteed that all of these will be uploaded every time. Rather, they are
+    /// only uploaded if already present on the device. To force the device to generate a specific
+    /// file it needs to be added to the `additional` section.
+    upload: &'static [&'static str],
+}
+
+fn create_data_request_response() -> DataRequestResponse {
+    DataRequestResponse {
+        parts: Parts {
+            upload: UPLOAD_PARTS,
+        },
+    }
+}
 
 fn validate_prosperodump(data: &[u8]) -> Result<(), BadStoreRequest> {
     if !data.starts_with(LZ4_MAGIC_HEADER) && !data.starts_with(PROSPERODUMP_MAGIC_HEADER) {
@@ -68,10 +106,14 @@ async fn extract_multipart(
 async fn handle(
     state: ServiceState,
     meta: RequestMeta,
-    _content_type: RawContentType,
+    content_type: RawContentType,
     request: Request,
 ) -> axum::response::Result<impl IntoResponse> {
-    // The crash dumps are transmitted as `...` in a multipart form-data/ request.
+    // Handle either a data request (send as json) or a crash dump (send as a multi-part)
+    if content_type.as_ref().contains(DATA_REQUEST_CONTENT_TYPE) {
+        return Ok(axum::Json(create_data_request_response()).into_response());
+    }
+
     let multipart = request.extract_with_state(&state).await?;
     let mut envelope = extract_multipart(multipart, meta, state.config()).await?;
     envelope.require_feature(Feature::PlaystationIngestion);
@@ -85,7 +127,7 @@ async fn handle(
     };
 
     // Return here needs to be a 200 with arbitrary text to make the sender happy.
-    Ok(TextResponse(id))
+    Ok(TextResponse(id).into_response())
 }
 
 pub fn route(config: &Config) -> MethodRouter<ServiceState> {